extract many things
